# openapi.yaml
# Generated by `python -m app.openapi_schema`. Do not edit by hand.
openapi: 3.1.0
info:
  title: last30days API
  summary: Private API for scored social/web evidence on a topic
  description: 'Async research jobs over the last30days engine. Authenticate with
    `Authorization: Bearer <key>`. Errors are `application/problem+json` with a stable
    `code`. Additive changes are allowed within /v1; breaking changes need /v2.'
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local / compose default
paths:
  /health:
    get:
      tags:
      - meta
      summary: Liveness (no auth)
      operationId: getHealth
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Gethealth
      security: []
  /v1/version:
    get:
      tags:
      - meta
      summary: Version
      operationId: getVersion
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionInfo'
        '401':
          description: Missing or invalid API key
          content:
            application/problem+json: &id001
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limit exceeded
          content:
            application/problem+json: *id001
          headers: &id002
            Retry-After:
              description: Seconds to wait
              schema:
                type: integer
      security:
      - HTTPBearer: []
  /v1/sources:
    get:
      tags:
      - meta
      summary: Sources
      operationId: listSources
      security:
      - HTTPBearer: []
      parameters:
      - name: refresh
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Refresh
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
        '401':
          content:
            application/problem+json: *id001
          description: Missing or invalid API key
        '429':
          content:
            application/problem+json: *id001
          description: Rate limit exceeded
          headers: *id002
        '502':
          content:
            application/problem+json: *id001
          description: Bad Gateway
        '422':
          description: Invalid request
          content:
            application/problem+json: *id001
  /v1/research:
    post:
      tags:
      - jobs
      summary: Start or reuse a research job
      operationId: createResearchJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRequest'
        required: true
      responses:
        '202':
          description: Accepted (new or reused job)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          headers:
            Location:
              description: URL of the job
              schema:
                type: string
        '401':
          description: Missing or invalid API key
          content:
            application/problem+json: *id001
        '429':
          description: Rate limit exceeded
          content:
            application/problem+json: *id001
          headers: *id002
        '422':
          description: Invalid request
          content:
            application/problem+json: *id001
        '503':
          description: Service Unavailable
          content:
            application/problem+json: *id001
          headers: *id002
      security:
      - HTTPBearer: []
  /v1/discover:
    post:
      tags:
      - jobs
      summary: Start or reuse a topic-less trending job
      operationId: createDiscoverJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoverRequest'
        required: true
      responses:
        '202':
          description: Accepted (new or reused job)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
          headers:
            Location:
              description: URL of the job
              schema:
                type: string
        '401':
          description: Missing or invalid API key
          content:
            application/problem+json: *id001
        '429':
          description: Rate limit exceeded
          content:
            application/problem+json: *id001
          headers: *id002
        '422':
          description: Invalid request
          content:
            application/problem+json: *id001
        '503':
          description: Service Unavailable
          content:
            application/problem+json: *id001
          headers: *id002
      security:
      - HTTPBearer: []
  /v1/jobs:
    get:
      tags:
      - jobs
      summary: List Jobs
      operationId: listJobs
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(queued|running|succeeded|partial|failed|canceled)$
          - type: 'null'
          title: Status
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: ^(research|discover)$
          - type: 'null'
          title: Kind
      - name: topic
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 300
          - type: 'null'
          title: Topic
      - name: client_ref
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Client Ref
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
        '401':
          content:
            application/problem+json: *id001
          description: Missing or invalid API key
        '429':
          content:
            application/problem+json: *id001
          description: Rate limit exceeded
          headers: *id002
        '422':
          content:
            application/problem+json: *id001
          description: Invalid request
  /v1/jobs/{job_id}:
    get:
      tags:
      - jobs
      summary: Get Job
      operationId: getJob
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          content:
            application/problem+json: *id001
          description: Missing or invalid API key
        '429':
          content:
            application/problem+json: *id001
          description: Rate limit exceeded
          headers: *id002
        '404':
          content:
            application/problem+json: *id001
          description: Not Found
        '422':
          description: Invalid request
          content:
            application/problem+json: *id001
    delete:
      tags:
      - jobs
      summary: Cancel (no-op if terminal)
      operationId: cancelJob
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          content:
            application/problem+json: *id001
          description: Missing or invalid API key
        '429':
          content:
            application/problem+json: *id001
          description: Rate limit exceeded
          headers: *id002
        '404':
          content:
            application/problem+json: *id001
          description: Not Found
        '422':
          description: Invalid request
          content:
            application/problem+json: *id001
  /v1/jobs/{job_id}/raw:
    get:
      tags:
      - jobs
      summary: Get Raw
      operationId: getJobRaw
      security:
      - HTTPBearer: []
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Unmodified upstream JSON (no compatibility guarantee)
          content:
            application/json:
              schema:
                type: object
        '401':
          content:
            application/problem+json: *id001
          description: Missing or invalid API key
        '429':
          content:
            application/problem+json: *id001
          description: Rate limit exceeded
          headers: *id002
        '404':
          content:
            application/problem+json: *id001
          description: Not Found
        '422':
          description: Invalid request
          content:
            application/problem+json: *id001
components:
  schemas:
    Cluster:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        item_ids:
          items:
            type: string
          type: array
          title: Item Ids
        sources:
          items:
            type: string
          type: array
          title: Sources
      type: object
      required:
      - id
      - title
      - item_ids
      - sources
      title: Cluster
    DiscoverRequest:
      properties:
        max_age_seconds:
          type: integer
          maximum: 2592000.0
          minimum: 0.0
          title: Max Age Seconds
          default: 21600
        force_refresh:
          type: boolean
          title: Force Refresh
          default: false
        timeout_seconds:
          type: integer
          maximum: 1800.0
          minimum: 30.0
          title: Timeout Seconds
          default: 900
        webhook_url:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Webhook Url
        client_ref:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Client Ref
        query:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Query
        limit:
          type: integer
          maximum: 25.0
          minimum: 1.0
          title: Limit
          default: 10
      additionalProperties: false
      type: object
      title: DiscoverRequest
    DiscoverResult:
      properties:
        topics:
          items:
            $ref: '#/components/schemas/DiscoverTopic'
          type: array
          title: Topics
        source_runs:
          items:
            $ref: '#/components/schemas/SourceRun'
          type: array
          title: Source Runs
        upstream:
          $ref: '#/components/schemas/UpstreamInfo'
      type: object
      required:
      - topics
      - source_runs
      - upstream
      title: DiscoverResult
    DiscoverTopic:
      properties:
        name:
          type: string
          title: Name
        score:
          type: number
          title: Score
        momentum:
          type: string
          title: Momentum
        sources:
          items:
            type: string
          type: array
          title: Sources
        suggested_query:
          type: string
          title: Suggested Query
      type: object
      required:
      - name
      - score
      - momentum
      - sources
      - suggested_query
      title: DiscoverTopic
    Engagement:
      properties:
        score:
          anyOf:
          - type: number
          - type: 'null'
          title: Score
        comments:
          anyOf:
          - type: number
          - type: 'null'
          title: Comments
        likes:
          anyOf:
          - type: number
          - type: 'null'
          title: Likes
        shares:
          anyOf:
          - type: number
          - type: 'null'
          title: Shares
        views:
          anyOf:
          - type: number
          - type: 'null'
          title: Views
      type: object
      title: Engagement
    EvidenceItem:
      properties:
        id:
          type: string
          title: Id
        source:
          type: string
          title: Source
        kind:
          type: string
          enum:
          - post
          - comment
          - video
          - transcript_excerpt
          - market
          - repo
          - article
          - other
          title: Kind
        url:
          type: string
          title: Url
        title:
          type: string
          title: Title
        text:
          type: string
          maxLength: 2000
          title: Text
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        published_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Published At
        engagement:
          $ref: '#/components/schemas/Engagement'
        relevance_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Relevance Score
        cluster_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
        extra:
          additionalProperties: true
          type: object
          title: Extra
          default: {}
      type: object
      required:
      - id
      - source
      - kind
      - url
      - title
      - text
      - engagement
      - relevance_score
      title: EvidenceItem
    Job:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          enum:
          - research
          - discover
          title: Kind
        status:
          type: string
          enum:
          - queued
          - running
          - succeeded
          - partial
          - failed
          - canceled
          title: Status
        request:
          additionalProperties: true
          type: object
          title: Request
        cache_hit:
          type: boolean
          title: Cache Hit
          default: false
        client_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Ref
        progress:
          $ref: '#/components/schemas/Progress'
          default:
            sources_total: 0
            sources_done: 0
        result:
          anyOf:
          - $ref: '#/components/schemas/ResearchResult'
          - $ref: '#/components/schemas/DiscoverResult'
          - type: 'null'
          title: Result
        error:
          anyOf:
          - $ref: '#/components/schemas/Problem'
          - type: 'null'
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
      type: object
      required:
      - id
      - kind
      - status
      - request
      - created_at
      title: Job
    JobList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/JobSummary'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - items
      title: JobList
    JobSummary:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          enum:
          - research
          - discover
          title: Kind
        status:
          type: string
          enum:
          - queued
          - running
          - succeeded
          - partial
          - failed
          - canceled
          title: Status
        request:
          additionalProperties: true
          type: object
          title: Request
        cache_hit:
          type: boolean
          title: Cache Hit
          default: false
        client_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Ref
        progress:
          $ref: '#/components/schemas/Progress'
          default:
            sources_total: 0
            sources_done: 0
        result:
          anyOf:
          - $ref: '#/components/schemas/ResearchResult'
          - $ref: '#/components/schemas/DiscoverResult'
          - type: 'null'
          title: Result
        error:
          anyOf:
          - $ref: '#/components/schemas/Problem'
          - type: 'null'
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
      type: object
      required:
      - id
      - kind
      - status
      - request
      - created_at
      title: JobSummary
      description: Same shape as Job; `result` is always omitted in listings.
    Market:
      properties:
        id:
          type: string
          title: Id
        question:
          type: string
          title: Question
        url:
          type: string
          title: Url
        probability:
          anyOf:
          - type: number
          - type: 'null'
          title: Probability
        volume_usd:
          anyOf:
          - type: number
          - type: 'null'
          title: Volume Usd
        closes_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Closes At
      type: object
      required:
      - id
      - question
      - url
      title: Market
    Problem:
      properties:
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          type: string
          title: Code
        detail:
          type: string
          title: Detail
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      title: Problem
    Progress:
      properties:
        sources_total:
          type: integer
          title: Sources Total
          default: 0
        sources_done:
          type: integer
          title: Sources Done
          default: 0
      type: object
      title: Progress
    ResearchRequest:
      properties:
        max_age_seconds:
          type: integer
          maximum: 2592000.0
          minimum: 0.0
          title: Max Age Seconds
          default: 21600
        force_refresh:
          type: boolean
          title: Force Refresh
          default: false
        timeout_seconds:
          type: integer
          maximum: 1800.0
          minimum: 30.0
          title: Timeout Seconds
          default: 900
        webhook_url:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Webhook Url
        client_ref:
          anyOf:
          - type: string
            maxLength: 200
          - type: 'null'
          title: Client Ref
        topic:
          type: string
          maxLength: 300
          minLength: 2
          title: Topic
        sources_include:
          items:
            type: string
          type: array
          title: Sources Include
          default: []
        sources_exclude:
          items:
            type: string
          type: array
          title: Sources Exclude
          default: []
      additionalProperties: false
      type: object
      required:
      - topic
      title: ResearchRequest
    ResearchResult:
      properties:
        topic:
          type: string
          title: Topic
        generated_at:
          type: string
          format: date-time
          title: Generated At
        window:
          $ref: '#/components/schemas/Window'
        evidence:
          items:
            $ref: '#/components/schemas/EvidenceItem'
          type: array
          title: Evidence
        clusters:
          items:
            $ref: '#/components/schemas/Cluster'
          type: array
          title: Clusters
        markets:
          items:
            $ref: '#/components/schemas/Market'
          type: array
          title: Markets
        source_runs:
          items:
            $ref: '#/components/schemas/SourceRun'
          type: array
          title: Source Runs
        upstream:
          $ref: '#/components/schemas/UpstreamInfo'
        raw_available:
          type: boolean
          title: Raw Available
          default: true
      type: object
      required:
      - topic
      - generated_at
      - window
      - evidence
      - clusters
      - markets
      - source_runs
      - upstream
      title: ResearchResult
    SourceList:
      properties:
        sources:
          items:
            $ref: '#/components/schemas/SourceStatus'
          type: array
          title: Sources
        checked_at:
          type: string
          format: date-time
          title: Checked At
        cached:
          type: boolean
          title: Cached
      type: object
      required:
      - sources
      - checked_at
      - cached
      title: SourceList
    SourceRun:
      properties:
        source:
          type: string
          title: Source
        status:
          type: string
          enum:
          - ok
          - empty
          - skipped
          - failed
          title: Status
        item_count:
          type: integer
          title: Item Count
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - source
      - status
      - item_count
      title: SourceRun
    SourceStatus:
      properties:
        name:
          type: string
          title: Name
        usable:
          type: boolean
          title: Usable
        state:
          type: string
          title: State
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
      type: object
      required:
      - name
      - usable
      - state
      title: SourceStatus
    UpstreamInfo:
      properties:
        repo:
          type: string
          title: Repo
        ref:
          type: string
          title: Ref
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
      type: object
      required:
      - repo
      - ref
      title: UpstreamInfo
    VersionInfo:
      properties:
        api_version:
          type: string
          title: Api Version
        schema_version:
          type: string
          title: Schema Version
        upstream_ref:
          type: string
          title: Upstream Ref
        upstream_repo:
          type: string
          title: Upstream Repo
      type: object
      required:
      - api_version
      - schema_version
      - upstream_ref
      - upstream_repo
      title: VersionInfo
    Window:
      properties:
        from:
          type: string
          format: date-time
          title: From
        to:
          type: string
          format: date-time
          title: To
      type: object
      required:
      - from
      - to
      title: Window
  securitySchemes:
    HTTPBearer:
      type: http
      description: Static API key from API_KEYS
      scheme: bearer